refactor env package to use private functions for handling environment reading. remove setting the logger inside the env package, this is not the responsability of the caller. Let env return environment variables only
