+++ internal/queue/google/pubsub.go
@@ -4,7 +4,7 @@ import (
     "sync"
     "context"
 
-    "github.com/jsirianni/relay/internal/env"
+    "github.com/jsirianni/relay/internal/util/env"
     "github.com/jsirianni/relay/internal/util/logger"
     "github.com/jsirianni/relay/internal/queue/qmessage"
 
@@ -34,7 +34,7 @@ func Init(topic string, l logger.Logger) (Google, error) {
         return p, errors.New("logger is not configured")
     }
 
-    p.ProjectID, err = env.ENVProjectID()
+    p.ProjectID, err = env.ProjectID()
     if err != nil {
         return p, err
     }
